Feat/favorites token

+++ src/utils/useFetchData.ts
@@ -7,8 +7,8 @@ import { ProductExpanded } from '../types';
 const BASE_URL = apiDBurl;
 
 //test data token
-const TOKEN =
-  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NCwidXNlcm5hbWUiOiJDaHJpc3RpbmEyIiwiaWF0IjoxNzE2ODIxNjU5LCJleHAiOjE3MTk0MTM2NTl9.tHE6xxlPtvgo59vmBlpTLyQZaBhr5pb_8ffU3HH98Ow';
+// const TOKEN =
+//   'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NCwidXNlcm5hbWUiOiJDaHJpc3RpbmEyIiwiaWF0IjoxNzE2ODIxNjU5LCJleHAiOjE3MTk0MTM2NTl9.tHE6xxlPtvgo59vmBlpTLyQZaBhr5pb_8ffU3HH98Ow';
 
 type FetchState<T> = {
   data: T[];
@@ -76,13 +76,22 @@ async function getProductsByNamespaceId(url: string, namespaceId: string) {
 
 //c
 async function addToFavorites(
-  userId: number,
   productId: number,
 ): Promise<void> {
   try {
-    if (!userId) {
-      userId = 1;
+    const token = localStorage.getItem('token')
+    const id = localStorage.getItem('userId');
+
+    if (id === null) {
+      throw new Error('userId is null');
     }
+
+    const userId = +id;
+
+    if(!token) {
+      throw new Error('Token is unavailable');
+    }
+
     const response = await axios.post(
       `${BASE_URL}users/favorites`,
       {
@@ -91,7 +100,7 @@ async function addToFavorites(
       },
       {
         headers: {
-          Authorization: `Bearer ${TOKEN}`,
+          Authorization: `Bearer ${token}`,
         },
       },
     );
@@ -102,14 +111,26 @@ async function addToFavorites(
 }
 
 async function removeFromFavorites(
-  userId: number,
   productId: number,
 ): Promise<void> {
   try {
+    const token = localStorage.getItem('token')
+    const id = localStorage.getItem('userId');
+
+    if (id === null) {
+      throw new Error('userId is unavailable');
+    }
+
+    const userId = +id;
+
+    if(!token) {
+      throw new Error('Token is unavailable');
+    }
+
     const response = await axios.delete(`${BASE_URL}users/favorites`, {
-      data: { userId, productId },
+      data: { userId: userId, productId },
       headers: {
-        Authorization: `Bearer ${TOKEN}`,
+        Authorization: `Bearer ${token}`,
       },
     });
     return response.data;
@@ -118,11 +139,24 @@ async function removeFromFavorites(
   }
 }
 
-async function getUserFavorites(userId: number) {
+async function getUserFavorites() {
   try {
+    const token = localStorage.getItem('token')
+    const id = localStorage.getItem('userId');
+
+    if (id === null) {
+      throw new Error('userId is unavailable');
+    }
+
+    const userId = +id;
+
+    if(!token) {
+      throw new Error('Token is unavailable');
+    }
+
     const response = await axios.get(`${BASE_URL}users/${userId}/favorites`, {
       headers: {
-        Authorization: `Bearer ${TOKEN}`,
+        Authorization: `Bearer ${token}`,
       },
     });
     return response.data;
@@ -130,8 +164,6 @@ async function getUserFavorites(userId: number) {
     throw new Error('Failed to fetch favorites');
   }
 }
-
-//c
 
 async function getDiscounts() {
   try {
@@ -142,7 +174,6 @@ async function getDiscounts() {
   }
 }
 
-//c
 async function getNewProducts() {
   try {
     const response = await axios.get(`${BASE_URL}products/new`);
@@ -154,11 +185,10 @@ async function getNewProducts() {
 
 //c
 async function isProductInFavorites(
-  userId: number,
   productId: number,
 ): Promise<boolean> {
   try {
-    const favorites = await getUserFavorites(userId);
+    const favorites = await getUserFavorites();
 
     return favorites.some(
       (favorite: Favorite) => favorite.productId === productId,
@@ -169,11 +199,10 @@ async function isProductInFavorites(
 }
 
 async function getOneFavorite(
-  userId: number,
   productId: number,
 ): Promise<Favorite | null> {
   try {
-    const favorites = await getUserFavorites(userId);
+    const favorites = await getUserFavorites();
     const favorite = favorites.find(
       (favorite: Favorite) => favorite.productId === productId,
     );
